requirements.txt
Django==3.2.16
django-split-settings==1.2.0
django-createsuperuserwithpassword==2.0.0
gunicorn==20.1.0
daphne==4.0.0
django-dotenv==1.4.2
psycopg2-binary==2.9.5
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.13.0
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.2
redis==4.4.0
django-celery-results==2.4.0
django-celery-beat==2.4.0
django-debug-toolbar==3.8.1
pyyaml==6.0
pygments==2.13.0
django-nested-inline==0.4.5
django-smart-selects==1.6.0
django-braces==1.15.0
django-import-export==3.0.2
pytest-django==4.5.2
pytest-html==3.2.0
allure-pytest==2.12.0
mysqlclient==2.1.1